Server error codes

+++ src/com/oblong/jelly/pool/PoolServerException.java
@@ -8,20 +8,20 @@ import com.oblong.jelly.Slaw;
 @Immutable
 public class PoolServerException extends PoolException {
 
-    public PoolServerException(int serverCode, String info) {
+    public PoolServerException(long serverCode, String info) {
         this(serverCode, null, info);
     }
 
-    public PoolServerException(int serverCode, Throwable cause) {
+    public PoolServerException(long serverCode, Throwable cause) {
         this(serverCode, null, cause);
     }
 
-    public PoolServerException(int serverCode, Slaw res, String info) {
+    public PoolServerException(long serverCode, Slaw res, String info) {
         super(serverCode, info);
         response = res;
     }
 
-    public PoolServerException(int serverCode, Slaw res, Throwable cause) {
+    public PoolServerException(long serverCode, Slaw res, Throwable cause) {
         super(serverCode, cause);
         response = res;
     }
